rename C.ulong to C.DWORD

+++ usb_winusb.go
@@ -54,7 +54,7 @@ const (
 	ERROR_BAD_COMMAND       C_DWORD = C.ERROR_BAD_COMMAND
 )
 
-func winusbError(err C.ulong) error {
+func winusbError(err C.DWORD) error {
 	if err != C.ERROR_SUCCESS {
 		return C_DWORD(err)
 	}
